requirements.txt
Flask==3.0.0
flask-cors==4.0.0
Werkzeug==3.0.1
langchain==0.1.0
langchain-community==0.0.10
langchain-google-genai==0.0.5
langchain-core==0.1.10
chromadb==0.4.22
python-dotenv==1.0.0
google-generativeai==0.3.2
PyPDF2==3.0.1
python-docx==1.1.0
docx2txt==0.8
requests==2.31.0 